refactor: react-query 적용 (professor)
