+++ apps/community/src/apis/member/professor/queries.ts
@@ -1,0 +1,14 @@
+import ProfessorService from './professor-service';
+
+const queryKeys = {
+  all: ['contacts'] as const,
+};
+
+const professorQueryOptions = {
+  all: () => ({
+    queryKey: queryKeys.all,
+    queryFn: () => ProfessorService.getProfessors(),
+  }),
+};
+
+export { professorQueryOptions };
