+++ apps/community/src/types/professor.ts
@@ -1,0 +1,10 @@
+interface Professor {
+  id: number;
+  name: string;
+  img?: string;
+  type: string;
+  contact: string;
+  email: string;
+}
+
+export type { Professor };
